+++ package.json
@@ -14,14 +14,11 @@
     "dev": "netlify dev"
   },
   "dependencies": {
-    "chrome-aws-lambda": "^3.1.1",
     "jsdom": "^16.2.2",
-    "node-fetch": "^2.6.0",
-    "puppeteer-core": "^3.2.0"
+    "node-fetch": "^2.6.0"
   },
   "devDependencies": {
     "netlify-cli": "^2.52.0",
-    "prettier": "^2.0.4",
-    "puppeteer": "^3.2.0"
+    "prettier": "^2.0.4"
   }
 }
